// prismicio-types.d.ts
interface NavDocumentData {
  /**
   * About field in *Nav*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: nav.about
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  about: prismic.KeyTextField;

  /**
   * Experience field in *Nav*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: nav.experience
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  experience: prismic.KeyTextField;

  /**
   * Projects field in *Nav*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: nav.projects
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  projects: prismic.KeyTextField;

  /**
   * Contact field in *Nav*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: nav.contact
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  contact: prismic.KeyTextField;
}